+++ internal/orchestrationhttp/task_handler.go
@@ -3,7 +3,6 @@ package orchestrationhttp
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"net/http"
@@ -16,7 +15,6 @@ import (
 	"github.com/johnjallday/ori-agent/internal/agentcomm"
 	orihttp "github.com/johnjallday/ori-agent/internal/http"
 	"github.com/johnjallday/ori-agent/internal/logger"
-	"github.com/johnjallday/ori-agent/internal/platform"
 	"github.com/johnjallday/ori-agent/internal/workspace"
 )
 
@@ -47,131 +45,6 @@ type FrontendScheduleConfigRaw struct {
 	CronExpr        string  `json:"cron_expr,omitempty"`
 	MaxRuns         int     `json:"max_runs,omitempty"`
 	EndDate         *string `json:"end_date,omitempty"`
-}
-
-// parseFlexibleTime parses a datetime string with or without timezone
-func parseFlexibleTime(s string) (*time.Time, error) {
-	if s == "" {
-		return nil, nil
-	}
-
-	// Try RFC3339 first (with timezone)
-	if t, err := time.Parse(time.RFC3339, s); err == nil {
-		return &t, nil
-	}
-
-	// Try without timezone (assume local)
-	if t, err := time.ParseInLocation("2006-01-02T15:04:05", s, time.Local); err == nil {
-		return &t, nil
-	}
-
-	// Try without seconds
-	if t, err := time.ParseInLocation("2006-01-02T15:04", s, time.Local); err == nil {
-		return &t, nil
-	}
-
-	return nil, fmt.Errorf("unable to parse time: %s", s)
-}
-
-// convertScheduleConfig converts frontend schedule format to backend format
-func convertScheduleConfig(raw json.RawMessage) *workspace.ScheduleConfig {
-	if raw == nil {
-		return nil
-	}
-
-	// First try parsing with raw strings to handle flexible datetime formats
-	var rawConfig FrontendScheduleConfigRaw
-	if err := json.Unmarshal(raw, &rawConfig); err != nil {
-		logger.Warn("Failed to parse schedule config", logger.Fields{"err": err})
-		return nil
-	}
-
-	config := &workspace.ScheduleConfig{
-		Type:     workspace.ScheduleType(rawConfig.Type),
-		MaxRuns:  rawConfig.MaxRuns,
-		CronExpr: rawConfig.CronExpr,
-	}
-
-	// Handle interval conversion (minutes to time.Duration)
-	if rawConfig.IntervalMinutes > 0 {
-		config.Interval = time.Duration(rawConfig.IntervalMinutes) * time.Minute
-		logger.Debug("Converted interval_minutes to Duration", logger.Fields{
-			"interval_minutes": rawConfig.IntervalMinutes,
-			"interval":         config.Interval,
-		})
-	}
-
-	// Handle time_of_day (frontend sends "time" or "time_of_day")
-	if rawConfig.Time != "" {
-		config.TimeOfDay = rawConfig.Time
-	} else if rawConfig.TimeOfDay != "" {
-		config.TimeOfDay = rawConfig.TimeOfDay
-	}
-
-	// Handle day_of_week
-	config.DayOfWeek = rawConfig.DayOfWeek
-
-	// Handle execute_at (frontend sends "run_at" or "execute_at") with flexible parsing
-	if rawConfig.RunAt != nil {
-		if t, err := parseFlexibleTime(*rawConfig.RunAt); err == nil {
-			config.ExecuteAt = t
-		} else {
-			logger.Warn("Failed to parse run_at time", logger.Fields{"value": *rawConfig.RunAt, "err": err})
-		}
-	} else if rawConfig.ExecuteAt != nil {
-		if t, err := parseFlexibleTime(*rawConfig.ExecuteAt); err == nil {
-			config.ExecuteAt = t
-		} else {
-			logger.Warn("Failed to parse execute_at time", logger.Fields{"value": *rawConfig.ExecuteAt, "err": err})
-		}
-	}
-
-	// Handle end_date with flexible parsing
-	if rawConfig.EndDate != nil {
-		if t, err := parseFlexibleTime(*rawConfig.EndDate); err == nil {
-			config.EndDate = t
-		} else {
-			logger.Warn("Failed to parse end_date time", logger.Fields{"value": *rawConfig.EndDate, "err": err})
-		}
-	}
-
-	logger.Debug("Converted frontend schedule to backend format", logger.Fields{
-		"type":        config.Type,
-		"interval":    config.Interval,
-		"time_of_day": config.TimeOfDay,
-		"day_of_week": config.DayOfWeek,
-		"execute_at":  config.ExecuteAt,
-	})
-
-	return config
-}
-
-func normalizeTaskSleepPolicy(policy string) string {
-	switch strings.ToLower(strings.TrimSpace(policy)) {
-	case "skip", "run_once_on_wake":
-		return strings.ToLower(strings.TrimSpace(policy))
-	default:
-		return "run_once_on_wake"
-	}
-}
-
-func normalizeWakeFallbackPolicy(policy string) string {
-	switch strings.ToLower(strings.TrimSpace(policy)) {
-	case "run_on_next_wake", "skip":
-		return strings.ToLower(strings.TrimSpace(policy))
-	default:
-		return "run_on_next_wake"
-	}
-}
-
-func normalizeWakeLeadMinutes(minutes int) int {
-	if minutes <= 0 {
-		return 5
-	}
-	if minutes > 120 {
-		return 120
-	}
-	return minutes
 }
 
 // TaskHandler manages task and scheduled task operations
@@ -737,84 +610,6 @@ func (th *TaskHandler) applyBasicFieldUpdates(task *workspace.Task, req *taskUpd
 		}
 		logger.Debug("Updated task kanban due date", logger.Fields{"task_id": req.TaskID, "kanban_due_date": val})
 	}
-}
-
-// applyScheduleUpdates applies schedule configuration changes to a task
-// Returns an error message if validation fails, empty string otherwise
-func (th *TaskHandler) applyScheduleUpdates(task *workspace.Task, req *taskUpdateRequest, schedule *workspace.ScheduleConfig, clearSchedule bool) string {
-	if schedule != nil {
-		task.Schedule = schedule
-		if task.ScheduleEnabled {
-			task.NextRun = th.calculateNextRun(task)
-		}
-		logger.Debug("Updated task schedule", logger.Fields{"task_id": req.TaskID})
-	}
-
-	if clearSchedule {
-		task.Schedule = nil
-		task.ScheduleEnabled = false
-		task.ScheduleName = ""
-		task.SleepPolicy = ""
-		task.WakeMacEnabled = false
-		task.WakeLeadMinutes = 0
-		task.WakeFallback = ""
-		task.NextRun = nil
-		logger.Debug("Cleared task schedule", logger.Fields{"task_id": req.TaskID})
-	}
-
-	if req.ScheduleEnabled != nil {
-		if *req.ScheduleEnabled {
-			taskTo := task.To
-			if req.To != nil {
-				taskTo = *req.To
-			}
-			if taskTo == "" || taskTo == "unassigned" {
-				return "Scheduled tasks must be assigned to an agent. Please assign an agent before enabling the schedule."
-			}
-		}
-
-		task.ScheduleEnabled = *req.ScheduleEnabled
-		if *req.ScheduleEnabled && task.Schedule != nil {
-			task.NextRun = th.calculateNextRun(task)
-		} else if !*req.ScheduleEnabled {
-			task.NextRun = nil
-		}
-		logger.Debug("Updated task schedule enabled", logger.Fields{"task_id": req.TaskID, "enabled": *req.ScheduleEnabled})
-	}
-
-	if req.ScheduleName != nil {
-		task.ScheduleName = *req.ScheduleName
-	}
-	if req.SleepPolicy != nil {
-		task.SleepPolicy = normalizeTaskSleepPolicy(*req.SleepPolicy)
-	}
-	if req.WakeMacEnabled != nil {
-		task.WakeMacEnabled = *req.WakeMacEnabled
-	}
-	if req.WakeLeadMinutes != nil {
-		task.WakeLeadMinutes = normalizeWakeLeadMinutes(*req.WakeLeadMinutes)
-	}
-	if req.WakeFallbackPolicy != nil {
-		task.WakeFallback = normalizeWakeFallbackPolicy(*req.WakeFallbackPolicy)
-	}
-
-	if task.WakeMacEnabled && (task.Schedule == nil || !task.ScheduleEnabled) {
-		task.WakeMacEnabled = false
-	}
-
-	return ""
-}
-
-// calculateNextRun calculates the next run time for a scheduled task
-func (th *TaskHandler) calculateNextRun(task *workspace.Task) *time.Time {
-	if task.Schedule == nil {
-		return nil
-	}
-	lastRun := time.Time{}
-	if task.LastRun != nil {
-		lastRun = *task.LastRun
-	}
-	return workspace.CalculateNextRun(*task.Schedule, lastRun)
 }
 
 // buildTaskUpdateEventData builds the event data for a task update
@@ -1482,724 +1277,12 @@ type SaveTaskResultRequest struct {
 	Format      string `json:"format,omitempty"`        // Optional: json, text, markdown, csv (default: text)
 }
 
-// handleSaveTaskResult handles POST /api/orchestration/tasks/{id}/save-result
-func (th *TaskHandler) handleSaveTaskResult(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	if r.Method != http.MethodPost {
-		orihttp.MethodNotAllowed(w)
-		return
-	}
-
-	var req SaveTaskResultRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		orihttp.BadRequest(w, "Invalid request body")
-		return
-	}
-
-	// Extract task ID from URL if not in body
-	if req.TaskID == "" {
-		pathParts := strings.Split(strings.TrimPrefix(r.URL.Path, "/api/orchestration/tasks/"), "/")
-		if len(pathParts) >= 1 && pathParts[0] != "" {
-			req.TaskID = pathParts[0]
-		}
-	}
-
-	if req.TaskID == "" {
-		orihttp.BadRequest(w, "Task ID is required")
-		return
-	}
-	if req.FilePath == "" {
-		orihttp.BadRequest(w, "File path is required")
-		return
-	}
-
-	// Security: Validate file path to prevent path traversal attacks
-	cleanFilePath := filepath.Clean(req.FilePath)
-	if strings.Contains(cleanFilePath, "..") {
-		orihttp.BadRequest(w, "Invalid file path: path traversal not allowed")
-		return
-	}
-	req.FilePath = cleanFilePath
-
-	// Set default format
-	if req.Format == "" {
-		req.Format = "text"
-	}
-
-	// Validate format
-	validFormats := map[string]bool{"json": true, "text": true, "markdown": true, "csv": true}
-	if !validFormats[req.Format] {
-		orihttp.BadRequest(w, "Format must be one of: json, text, markdown, csv")
-		return
-	}
-
-	// Find the task
-	task, ws, err := th.getTaskWithWorkspace(req.TaskID)
-	if err != nil {
-		orihttp.NotFound(w, fmt.Sprintf("Task not found: %v", err))
-		return
-	}
-
-	if task.Result == "" {
-		orihttp.BadRequest(w, "Task has no result to save")
-		return
-	}
-
-	var finalPath string
-
-	if req.StoreNodeID != "" {
-		// Save via store node
-		var storeNode *workspace.StoreNode
-		for i := range ws.StoreNodes {
-			if ws.StoreNodes[i].ID == req.StoreNodeID || ws.StoreNodes[i].CanvasNodeID == req.StoreNodeID {
-				storeNode = &ws.StoreNodes[i]
-				break
-			}
-		}
-
-		if storeNode == nil {
-			orihttp.NotFound(w, "Store node not found")
-			return
-		}
-
-		// Override format with store node's format
-		storeNode.Format = req.Format
-
-		dataToStore := task.Result
-		if req.Format == "csv" {
-			dataToStore = workspace.TaskResultToCSV(task, task.Result, time.Now().Format("20060102-150405"), "")
-		}
-
-		// Write to store
-		if err := workspace.WriteToStoreForWorkspace(storeNode, th.workspaceStore, ws.ID, req.FilePath, dataToStore); err != nil {
-			orihttp.InternalError(w, fmt.Sprintf("Failed to save result: %v", err))
-			return
-		}
-
-		finalPath, _ = workspace.BuildFinalStorePath(storeNode, th.workspaceStore, ws.ID, req.FilePath)
-
-		// Save workspace to persist store node stats
-		if err := th.workspaceStore.Save(ws); err != nil {
-			logger.Warn("Failed to save workspace after store write", logger.Fields{"error": err})
-		}
-	} else {
-		// Direct file save (for Quick Save or custom path)
-		// Format data based on format type
-		var formattedData []byte
-		switch req.Format {
-		case "json":
-			// Pretty-print JSON
-			var obj any
-			if err := json.Unmarshal([]byte(task.Result), &obj); err != nil {
-				// If not valid JSON, treat as plain text
-				formattedData = []byte(task.Result)
-			} else {
-				formattedData, _ = json.MarshalIndent(obj, "", "  ")
-			}
-		case "csv":
-			formattedData = []byte(workspace.TaskResultToCSV(task, task.Result, time.Now().Format("20060102-150405"), ""))
-		default:
-			formattedData = []byte(task.Result)
-		}
-
-		// Create directories
-		dir := filepath.Dir(req.FilePath)
-		if err := os.MkdirAll(dir, 0755); err != nil {
-			orihttp.InternalError(w, fmt.Sprintf("Failed to create directories: %v", err))
-			return
-		}
-
-		// Write file
-		if err := os.WriteFile(req.FilePath, formattedData, 0644); err != nil {
-			orihttp.InternalError(w, fmt.Sprintf("Failed to write file: %v", err))
-			return
-		}
-
-		finalPath = req.FilePath
-	}
-
-	logger.Info("Saved task result", logger.Fields{
-		"task_id":   req.TaskID,
-		"file_path": finalPath,
-		"format":    req.Format,
-	})
-
-	w.WriteHeader(http.StatusOK)
-	orihttp.WriteJSON(w, map[string]any{
-		"success":   true,
-		"message":   "Result saved successfully",
-		"file_path": finalPath,
-		"task_id":   req.TaskID,
-	})
-}
-
 type taskOutputReviewRequest struct {
 	Action        string   `json:"action"`
 	HistoryIndex  *int     `json:"history_index"`
 	Result        string   `json:"result,omitempty"`
 	ApprovedBy    string   `json:"approved_by,omitempty"`
 	TargetColumns []string `json:"target_columns,omitempty"`
-}
-
-func (th *TaskHandler) handleTaskOutputReview(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-
-	if r.Method != http.MethodPost {
-		orihttp.MethodNotAllowed(w)
-		return
-	}
-
-	pathParts := strings.Split(strings.TrimPrefix(r.URL.Path, "/api/orchestration/tasks/"), "/")
-	if len(pathParts) < 2 || pathParts[0] == "" {
-		orihttp.BadRequest(w, "task_id is required in URL path")
-		return
-	}
-	taskID := pathParts[0]
-
-	var req taskOutputReviewRequest
-	if !orihttp.ParseJSONBody(w, r, &req) {
-		return
-	}
-	action := strings.ToLower(strings.TrimSpace(req.Action))
-	if action == "" {
-		orihttp.BadRequest(w, "action is required")
-		return
-	}
-
-	task, ws, err := th.getTaskWithWorkspace(taskID)
-	if err != nil {
-		orihttp.RespondErrorWithErr(w, http.StatusNotFound, "Task not found", err)
-		return
-	}
-	historyIndex := resolveTaskReviewHistoryIndex(task, req.HistoryIndex)
-	if historyIndex < 0 || historyIndex >= len(task.ExecutionHistory) {
-		orihttp.BadRequest(w, "history_index does not identify a reviewable run")
-		return
-	}
-
-	// Set by the cases that attempt to store a row so the shared response
-	// below can report whether the row was actually written. A CSV header
-	// mismatch holds the row for review (StorageStatus=skipped_invalid) while
-	// the action itself succeeds, so "success" alone is not enough for a
-	// non-UI client to tell that nothing was appended.
-	var reviewValidation *workspace.TaskValidationResult
-
-	switch action {
-	case "inspect", "copy_raw":
-		th.publishOutputContractReviewEvent(ws.ID, task.ID, action, task.ExecutionHistory[historyIndex].Validation)
-		entry := task.ExecutionHistory[historyIndex]
-		w.WriteHeader(http.StatusOK)
-		orihttp.WriteJSON(w, map[string]any{
-			"success":           true,
-			"task_id":           task.ID,
-			"history_index":     historyIndex,
-			"result":            entry.Result,
-			"summary":           entry.Summary,
-			"validation_result": entry.Validation,
-		})
-		return
-	case "retry_normalization":
-		entry := task.ExecutionHistory[historyIndex]
-		rawResult := strings.TrimSpace(entry.Result)
-		if rawResult == "" {
-			rawResult = strings.TrimSpace(entry.Summary)
-		}
-		if rawResult == "" {
-			rawResult = strings.TrimSpace(task.Result)
-		}
-		if rawResult == "" {
-			orihttp.BadRequest(w, "raw result is required to retry normalization")
-			return
-		}
-		reviewTask := taskOutputReviewValidationTaskForEntry(task, entry)
-		if reviewTask == nil || reviewTask.OutputSpec == nil {
-			orihttp.BadRequest(w, "retry_normalization requires a structured output spec snapshot")
-			return
-		}
-		var assistant workspace.TaskOutputSpecAssistant
-		if candidate, ok := th.taskHandler.(workspace.TaskOutputSpecAssistant); ok {
-			assistant = candidate
-		}
-		validation, csvData := workspace.ValidateTaskOutputSpecResultWithAssistant(r.Context(), reviewTask, rawResult, assistant)
-		if validation.ValidationStatus == workspace.TaskValidationPassed {
-			if err := appendApprovedTaskCSV(th.workspaceStore, ws, task, csvData); err != nil {
-				if !recordTaskOutputReviewCSVHeaderMismatch(validation, err) {
-					orihttp.InternalError(w, fmt.Sprintf("Failed to append retried normalization: %v", err))
-					return
-				}
-			} else {
-				validation.StorageStatus = workspace.TaskStorageAppended
-			}
-		}
-		now := time.Now().UTC()
-		validation.ValidatedAt = &now
-		if err := th.recordTaskOutputReviewValidation(ws.ID, taskID, historyIndex, action, validation); err != nil {
-			orihttp.InternalError(w, fmt.Sprintf("Failed to record normalization retry: %v", err))
-			return
-		}
-		reviewValidation = validation
-	case "rerun":
-		if err := th.startTaskOutputReviewRerun(ws.ID, task.ID); err != nil {
-			orihttp.RespondErrorWithErr(w, http.StatusBadRequest, "Failed to re-run task", err)
-			return
-		}
-		th.publishOutputContractReviewEvent(ws.ID, task.ID, action, task.ExecutionHistory[historyIndex].Validation)
-		w.WriteHeader(http.StatusAccepted)
-		orihttp.WriteJSON(w, map[string]any{
-			"success": true,
-			"message": "Task re-run started",
-			"task_id": task.ID,
-		})
-		return
-	case "dismiss":
-		now := time.Now().UTC()
-		if err := th.workspaceStore.Update(ws.ID, func(fresh *workspace.Workspace) error {
-			return fresh.MutateTask(taskID, func(t *workspace.Task) error {
-				if historyIndex < 0 || historyIndex >= len(t.ExecutionHistory) {
-					return fmt.Errorf("history entry no longer exists")
-				}
-				validation := t.ExecutionHistory[historyIndex].Validation
-				if validation == nil {
-					validation = &workspace.TaskValidationResult{}
-				}
-				validation.ValidationStatus = workspace.TaskValidationDismissed
-				if validation.StorageStatus == "" {
-					validation.StorageStatus = workspace.TaskStorageSkippedInvalid
-				}
-				validation.ValidatedAt = &now
-				t.ExecutionHistory[historyIndex].Validation = validation
-				workspace.MirrorTaskValidationResult(fresh.ID, t.ID, t.ExecutionHistory[historyIndex].RunID, validation)
-				th.publishOutputContractReviewEvent(fresh.ID, t.ID, action, validation)
-				return nil
-			})
-		}); err != nil {
-			orihttp.InternalError(w, fmt.Sprintf("Failed to dismiss review: %v", err))
-			return
-		}
-	case "approve_append":
-		draft := strings.TrimSpace(req.Result)
-		if draft == "" {
-			draft = strings.TrimSpace(task.ExecutionHistory[historyIndex].Result)
-		}
-		if draft == "" {
-			orihttp.BadRequest(w, "result is required for manual approval")
-			return
-		}
-		reviewTask := taskOutputReviewValidationTask(task, task.ExecutionHistory[historyIndex].Validation)
-		validation, csvData := validateTaskOutputReviewApproval(reviewTask, draft)
-		if validation.ValidationStatus != workspace.TaskValidationPassed {
-			th.publishOutputContractReviewEvent(ws.ID, task.ID, action, validation)
-			w.WriteHeader(http.StatusBadRequest)
-			orihttp.WriteJSON(w, map[string]any{
-				"success":           false,
-				"validation_result": validation,
-				"message":           "Edited result does not match the output contract.",
-			})
-			return
-		}
-		if err := appendApprovedTaskCSV(th.workspaceStore, ws, task, csvData); err != nil {
-			if recordTaskOutputReviewCSVHeaderMismatch(validation, err) {
-				now := time.Now().UTC()
-				validation.ValidatedAt = &now
-				if err := th.recordTaskOutputReviewValidation(ws.ID, taskID, historyIndex, action, validation); err != nil {
-					orihttp.InternalError(w, fmt.Sprintf("Failed to record append review: %v", err))
-					return
-				}
-				reviewValidation = validation
-				break
-			}
-			orihttp.InternalError(w, fmt.Sprintf("Failed to append approved result: %v", err))
-			return
-		}
-		now := time.Now().UTC()
-		validation.ValidationStatus = workspace.TaskValidationManuallyApproved
-		validation.StorageStatus = workspace.TaskStorageManuallyAppended
-		validation.Errors = nil
-		validation.ManualApproval = &workspace.TaskManualApproval{
-			ApprovedAt: now,
-			ApprovedBy: strings.TrimSpace(req.ApprovedBy),
-		}
-		validation.ValidatedAt = &now
-		if err := th.workspaceStore.Update(ws.ID, func(fresh *workspace.Workspace) error {
-			return fresh.MutateTask(taskID, func(t *workspace.Task) error {
-				if historyIndex < 0 || historyIndex >= len(t.ExecutionHistory) {
-					return fmt.Errorf("history entry no longer exists")
-				}
-				t.ExecutionHistory[historyIndex].Validation = validation
-				workspace.MirrorTaskValidationResult(fresh.ID, t.ID, t.ExecutionHistory[historyIndex].RunID, validation)
-				th.publishOutputContractReviewEvent(fresh.ID, t.ID, action, validation)
-				return nil
-			})
-		}); err != nil {
-			orihttp.InternalError(w, fmt.Sprintf("Failed to record approval: %v", err))
-			return
-		}
-		reviewValidation = validation
-	case "reproject_to_destination":
-		entry := task.ExecutionHistory[historyIndex]
-		rawResult := strings.TrimSpace(entry.Result)
-		if rawResult == "" {
-			rawResult = strings.TrimSpace(entry.Summary)
-		}
-		if rawResult == "" {
-			rawResult = strings.TrimSpace(task.Result)
-		}
-		if rawResult == "" {
-			orihttp.BadRequest(w, "raw result is required to reproject")
-			return
-		}
-		targetColumns := req.TargetColumns
-		if len(targetColumns) == 0 {
-			targetColumns = expectedColumnsFromValidation(entry.Validation)
-		}
-		if len(targetColumns) == 0 {
-			orihttp.BadRequest(w, "no destination columns available to match; provide target_columns")
-			return
-		}
-		var assistant workspace.TaskOutputSpecAssistant
-		if candidate, ok := th.taskHandler.(workspace.TaskOutputSpecAssistant); ok {
-			assistant = candidate
-		}
-		csvData, _, reprojErr := workspace.ReprojectResultToColumns(r.Context(), task, rawResult, targetColumns, assistant)
-		if reprojErr != nil {
-			orihttp.InternalError(w, fmt.Sprintf("Failed to reorganize result: %v", reprojErr))
-			return
-		}
-		validation := entry.Validation
-		if validation == nil {
-			validation = &workspace.TaskValidationResult{}
-		}
-		if err := appendApprovedTaskCSV(th.workspaceStore, ws, task, csvData); err != nil {
-			// A mismatch here would be unexpected (we matched the file's header),
-			// but surface it for review rather than failing silently.
-			if recordTaskOutputReviewCSVHeaderMismatch(validation, err) {
-				now := time.Now().UTC()
-				validation.ValidatedAt = &now
-				if recordErr := th.recordTaskOutputReviewValidation(ws.ID, taskID, historyIndex, action, validation); recordErr != nil {
-					orihttp.InternalError(w, fmt.Sprintf("Failed to record reorganize review: %v", recordErr))
-					return
-				}
-				reviewValidation = validation
-				break
-			}
-			orihttp.InternalError(w, fmt.Sprintf("Failed to append reorganized result: %v", err))
-			return
-		}
-		now := time.Now().UTC()
-		validation.ValidationStatus = workspace.TaskValidationManuallyApproved
-		validation.StorageStatus = workspace.TaskStorageManuallyAppended
-		validation.Errors = nil
-		validation.ManualApproval = &workspace.TaskManualApproval{
-			ApprovedAt: now,
-			ApprovedBy: strings.TrimSpace(req.ApprovedBy),
-		}
-		validation.ValidatedAt = &now
-		if err := th.recordTaskOutputReviewValidation(ws.ID, taskID, historyIndex, action, validation); err != nil {
-			orihttp.InternalError(w, fmt.Sprintf("Failed to record reorganized append: %v", err))
-			return
-		}
-		reviewValidation = validation
-	default:
-		orihttp.BadRequest(w, "action must be inspect, copy_raw, dismiss, rerun, retry_normalization, reproject_to_destination, or approve_append")
-		return
-	}
-
-	updatedTask, err := th.communicator.GetTask(taskID)
-	if err != nil {
-		orihttp.InternalError(w, fmt.Sprintf("Failed to load updated task: %v", err))
-		return
-	}
-	resp := map[string]any{
-		"success": true,
-		"task":    updatedTask,
-	}
-	// When the action attempted to store a row, report whether it actually
-	// landed. A CSV header mismatch holds the row for review rather than
-	// appending it, so "success" (the action was processed) must not be read
-	// as "the row was stored".
-	if reviewValidation != nil {
-		resp["validation_status"] = reviewValidation.ValidationStatus
-		resp["storage_status"] = reviewValidation.StorageStatus
-		resp["stored"] = reviewValidation.StorageStatus == workspace.TaskStorageAppended ||
-			reviewValidation.StorageStatus == workspace.TaskStorageSaved ||
-			reviewValidation.StorageStatus == workspace.TaskStorageManuallyAppended
-	}
-	w.WriteHeader(http.StatusOK)
-	orihttp.WriteJSON(w, resp)
-}
-
-func (th *TaskHandler) publishOutputContractReviewEvent(workspaceID, taskID, action string, validation *workspace.TaskValidationResult) {
-	if th.eventBus == nil {
-		return
-	}
-	data := map[string]any{
-		"task_id": taskID,
-		"action":  "review_action",
-		"review":  strings.TrimSpace(action),
-	}
-	if validation != nil {
-		data["validation_status"] = validation.ValidationStatus
-		data["storage_status"] = validation.StorageStatus
-		data["contract_version"] = validation.ContractVersion
-		data["error_count"] = len(validation.Errors)
-	}
-	th.eventBus.Publish(workspace.NewWorkspaceEvent(workspace.EventTaskOutput, workspaceID, "task.output_contract", data))
-}
-
-func (th *TaskHandler) recordTaskOutputReviewValidation(workspaceID, taskID string, historyIndex int, action string, validation *workspace.TaskValidationResult) error {
-	return th.workspaceStore.Update(workspaceID, func(fresh *workspace.Workspace) error {
-		return fresh.MutateTask(taskID, func(t *workspace.Task) error {
-			if historyIndex < 0 || historyIndex >= len(t.ExecutionHistory) {
-				return fmt.Errorf("history entry no longer exists")
-			}
-			t.ExecutionHistory[historyIndex].Validation = validation
-			workspace.MirrorTaskValidationResult(fresh.ID, t.ID, t.ExecutionHistory[historyIndex].RunID, validation)
-			th.publishOutputContractReviewEvent(fresh.ID, t.ID, action, validation)
-			return nil
-		})
-	})
-}
-
-// expectedColumnsFromValidation pulls the destination file's column header out
-// of a recorded csv_header_mismatch error, so a reproject can target the exact
-// columns the existing CSV expects.
-func expectedColumnsFromValidation(validation *workspace.TaskValidationResult) []string {
-	if validation == nil {
-		return nil
-	}
-	for _, e := range validation.Errors {
-		if strings.EqualFold(strings.TrimSpace(e.Code), "csv_header_mismatch") && len(e.Expected) > 0 {
-			return append([]string(nil), e.Expected...)
-		}
-	}
-	return nil
-}
-
-func recordTaskOutputReviewCSVHeaderMismatch(validation *workspace.TaskValidationResult, err error) bool {
-	var mismatch *workspace.CSVHeaderMismatchError
-	if !errors.As(err, &mismatch) {
-		return false
-	}
-	if validation == nil {
-		return false
-	}
-	validation.ValidationStatus = workspace.TaskValidationNeedsReview
-	validation.StorageStatus = workspace.TaskStorageSkippedInvalid
-	validation.Errors = append(validation.Errors, workspace.TaskValidationError{
-		Code:     "csv_header_mismatch",
-		Message:  mismatch.Error(),
-		Expected: append([]string(nil), mismatch.Expected...),
-		Actual:   append([]string(nil), mismatch.Actual...),
-	})
-	return true
-}
-
-func validateTaskOutputReviewApproval(task *workspace.Task, result string) (*workspace.TaskValidationResult, string) {
-	if task != nil && task.OutputSpec != nil {
-		return workspace.ValidateTaskOutputSpecResult(task, result)
-	}
-	return workspace.ValidateTaskOutputContractResult(task, result)
-}
-
-func taskOutputReviewValidationTask(task *workspace.Task, validation *workspace.TaskValidationResult) *workspace.Task {
-	if task == nil {
-		return nil
-	}
-	reviewTask := *task
-	if validation != nil && validation.OutputSpec != nil {
-		reviewTask.OutputSpec = workspace.SnapshotTaskOutputSpec(validation.OutputSpec)
-		reviewTask.OutputSchema = nil
-		reviewTask.OutputContract = nil
-		if reviewTask.OutputSpec != nil {
-			reviewTask.OutputSchema = reviewTask.OutputSpec.Schema
-			reviewTask.OutputContract = reviewTask.OutputSpec.Contract
-		}
-	}
-	return &reviewTask
-}
-
-func taskOutputReviewValidationTaskForEntry(task *workspace.Task, entry workspace.TaskExecution) *workspace.Task {
-	reviewTask := taskOutputReviewValidationTask(task, entry.Validation)
-	if reviewTask == nil {
-		return nil
-	}
-	reviewTask.CurrentRunID = strings.TrimSpace(entry.RunID)
-	reviewTask.ExecutionHistory = []workspace.TaskExecution{entry}
-	if reviewTask.Context == nil {
-		reviewTask.Context = map[string]any{}
-	}
-	return reviewTask
-}
-
-func resolveTaskReviewHistoryIndex(task *workspace.Task, requested *int) int {
-	if task == nil || len(task.ExecutionHistory) == 0 {
-		return -1
-	}
-	if requested != nil && *requested >= 0 && *requested < len(task.ExecutionHistory) {
-		return *requested
-	}
-	for i := len(task.ExecutionHistory) - 1; i >= 0; i-- {
-		validation := task.ExecutionHistory[i].Validation
-		if validation != nil && validation.ValidationStatus == workspace.TaskValidationNeedsReview {
-			return i
-		}
-	}
-	return -1
-}
-
-func (th *TaskHandler) startTaskOutputReviewRerun(workspaceID, taskID string) error {
-	if th.taskHandler == nil {
-		return fmt.Errorf("task execution not available")
-	}
-
-	ws, err := th.workspaceStore.Get(workspaceID)
-	if err != nil {
-		return err
-	}
-	task, err := ws.GetTask(taskID)
-	if err != nil {
-		return err
-	}
-	if task.Status == workspace.TaskStatusInProgress {
-		return fmt.Errorf("task is already in progress")
-	}
-
-	subtasks := ws.GetSubtasks(task.ID)
-	if len(subtasks) > 0 {
-		for _, subtask := range subtasks {
-			if subtask.Status == workspace.TaskStatusInProgress {
-				return fmt.Errorf("a subtask is already in progress")
-			}
-			if subtask.To == "" || subtask.To == "unassigned" {
-				return fmt.Errorf("all subtasks must be assigned to an agent before execution")
-			}
-		}
-		go th.executeParentTaskSequence(ws.ID, task.ID)
-		return nil
-	}
-
-	if err := task.SetStatus(workspace.TaskStatusPending); err != nil {
-		return err
-	}
-	workspace.ResetTaskRuntime(task)
-	if err := ws.UpdateTask(*task); err != nil {
-		return err
-	}
-	if err := th.workspaceStore.Save(ws); err != nil {
-		return err
-	}
-
-	go func() {
-		fresh, err := th.workspaceStore.Get(workspaceID)
-		if err != nil {
-			logger.Error("Failed to reload workspace for review re-run", logger.Fields{"workspace_id": workspaceID, "error": err})
-			return
-		}
-		rerunTask, err := fresh.GetTask(taskID)
-		if err != nil {
-			logger.Error("Task not found for review re-run", logger.Fields{"task_id": taskID, "error": err})
-			return
-		}
-		if _, err := th.executeTaskWithDependencies(fresh, rerunTask); err != nil {
-			var blockedErr *workspace.TaskBlockedError
-			if errors.As(err, &blockedErr) {
-				return
-			}
-			logger.Error("Review task re-run failed", logger.Fields{"task_id": taskID, "error": err})
-		}
-	}()
-	return nil
-}
-
-func appendApprovedTaskCSV(store workspace.Store, ws *workspace.Workspace, task *workspace.Task, csvData string) error {
-	if ws == nil || task == nil {
-		return fmt.Errorf("workspace and task are required")
-	}
-	storage := task.ResultStorage
-	if storage == nil || !storage.Enabled {
-		return fmt.Errorf("task result storage is not enabled")
-	}
-	if strings.ToLower(strings.TrimSpace(storage.WriteMode)) != "append" {
-		return fmt.Errorf("manual approval append requires append-to-CSV storage")
-	}
-	if strings.TrimSpace(csvData) == "" {
-		return fmt.Errorf("approved CSV is empty")
-	}
-
-	// The dataset is JSONL; convert the approved CSV to JSONL records and append
-	// those so the approve-held-run path writes the same .jsonl as the executor.
-	jsonlData, err := workspace.CSVToJSONL(csvData)
-	if err != nil {
-		return err
-	}
-
-	storeFilePath := storage.FilePath
-	if strings.TrimSpace(storeFilePath) == "" {
-		storeFilePath = workspace.AppendJSONLFileName(task, storage)
-	}
-
-	if strings.TrimSpace(storage.StoreNodeID) != "" {
-		var storeNode *workspace.StoreNode
-		for i := range ws.StoreNodes {
-			if ws.StoreNodes[i].ID == storage.StoreNodeID || ws.StoreNodes[i].CanvasNodeID == storage.StoreNodeID {
-				storeNode = &ws.StoreNodes[i]
-				break
-			}
-		}
-		if storeNode == nil {
-			return fmt.Errorf("store node %q not found", storage.StoreNodeID)
-		}
-		storeNodeCopy := *storeNode
-		storeNodeCopy.WriteMode = "append"
-		storeNodeCopy.Format = "jsonl"
-		// JSONL records are self-describing; append them directly (no CSV
-		// header reconciliation).
-		if err := workspace.WriteToStoreForWorkspace(&storeNodeCopy, store, ws.ID, storeFilePath, jsonlData); err != nil {
-			return err
-		}
-		storeNode.LastWriteTime = storeNodeCopy.LastWriteTime
-		storeNode.WriteCount = storeNodeCopy.WriteCount
-		storeNode.LastFilePath = storeNodeCopy.LastFilePath
-		storeNode.LastError = storeNodeCopy.LastError
-		storeNode.UpdatedAt = storeNodeCopy.UpdatedAt
-		return nil
-	}
-
-	filePath := storage.FilePath
-	if workspace.ResultStorageUsesWorkspaceFolder(storage) {
-		baseDir, _, err := workspace.ResolveWorkspaceFolderBaseDir(store, ws.ID, storage.WorkspaceFolder)
-		if err != nil {
-			return err
-		}
-		relativeFilePath := strings.TrimSpace(filePath)
-		if relativeFilePath == "" {
-			relativeFilePath = storeFilePath
-		} else if strings.HasSuffix(relativeFilePath, "/") || !strings.Contains(filepath.Base(relativeFilePath), ".") {
-			relativeFilePath = filepath.Join(relativeFilePath, workspace.AppendJSONLFileName(task, storage))
-		}
-		finalPath, err := workspace.BuildFinalPath(baseDir, relativeFilePath)
-		if err != nil {
-			return err
-		}
-		filePath = finalPath
-	} else if strings.TrimSpace(filePath) == "" {
-		baseOutputDir := ""
-		if store != nil {
-			baseOutputDir = store.GetOutputsPath(ws.ID)
-		}
-		if baseOutputDir == "" {
-			fallback, err := platform.GetDefaultOutputDir()
-			if err != nil {
-				fallback = "outputs"
-			}
-			baseOutputDir = filepath.Join(fallback, ws.Name)
-		}
-		filePath = filepath.Join(baseOutputDir, storeFilePath)
-	} else if strings.HasSuffix(filePath, "/") || !strings.Contains(filepath.Base(filePath), ".") {
-		filePath = filepath.Join(filePath, workspace.AppendJSONLFileName(task, storage))
-	}
-	return workspace.AppendJSONLToFile(filePath, jsonlData)
 }
 
 // BulkDeleteTasksHandler handles DELETE /api/orchestration/tasks/bulk
